+++ src/index.ts
@@ -18,3 +18,4 @@ export * from './maybe';
 export * from './list';
 export * from './reviver';
 export * from './lazy';
+export * from './stream';